+++ Assets/Scripts/SideBorderCollision.cs
@@ -10,6 +10,9 @@ public class SideBorderCollision : MonoBehaviour
     
     // Which side this border is on
     [SerializeField] private int side;
+    
+    // The audio to play on the border score hit
+    [SerializeField] private AudioClip hitAudio;
 
 
     void Awake()
@@ -19,12 +22,15 @@ public class SideBorderCollision : MonoBehaviour
         _forcefield = GetComponent<Forcefield>();
     }
     
+    /* Triggered when this border is passed by the ball */
     void OnCollisionEnter2D(Collision2D collision)
     {
 
         if (!_forcefield.IsActive)
         {
             _game.OnBorderHit(side);
+            
+            GlobalAudio.Instance.PlayAudio(hitAudio);
         }
     }
 }